added the precision calculation mechanism

+++ trig_simd.h
@@ -65,6 +65,6 @@
 
 #endif
 
-void sin_simd(double *input, double *res, size_t n, int prec);
+void sin_simd(double *input, double *res, size_t n, double prec);
 void tan_simd(double *input, double *res, size_t n, int prec);
 #endif
